Prevent NPE in finalizer thread (which leaves temporary files behind) in FileBackedList large enough to have a FileBackedLongList, when the garbage collector reclaims the latter before the former. Related tests, and related improvements to be more defensive in finalize() and close(). Give PdfBoxDocument a pool of AutoCloseable instances that need cleanup when the document is closed.

+++ src/org/lockss/util/CountingRandomAccessFile.java
@@ -351,14 +351,10 @@ public class CountingRandomAccessFile extends RandomAccessFile {
       super.close();
     }
     finally {
-      try {
-        if (deleteFile) {
-          file.delete();
-        }
+      if (deleteFile && file != null) {
+        file.delete();
       }
-      catch (Throwable thr) {
-        // ignore
-      }
+      file = null;
     }
   }
   
@@ -560,7 +556,7 @@ public class CountingRandomAccessFile extends RandomAccessFile {
    * @see https://www.programcreek.com/java-api-examples/index.php?source_dir=cp-common-utils-master/core/main/src/com/complexible/common/io/MMapUtil.java
    */
   public static boolean unmap(final MappedByteBuffer mbbuf) {
-    if (unmapSupported) {
+    if (unmapSupported && mbbuf != null) {
       try { 
         AccessController.doPrivileged(new PrivilegedExceptionAction<Object>() { 
           public Object run() throws Exception { 
@@ -579,7 +575,7 @@ public class CountingRandomAccessFile extends RandomAccessFile {
         // ignore
       } 
     }
-    return false;
+    return mbbuf == null;
   }
 
 }
